Setting up CI for Redis
This commit sets up CI system for Redis.

For local testing you can use vagrant to launch a VM and use molecule
to test the playbook.

For travis-ci you can use aws to spin up a vm.

Docker will be removed at some point as there are issues with systemd
and it does not make sense to try to resolve them for the limited use
case of testing (Vagrant / EC2 work well for this purpose).